Update CHANGELOG.md
Revised order of the changelog: any New Features then any Changes then any Bug Fixes.
Updated the Support Statement to reflect our latest guidance. We now point to our guide to determine what is supported. We have a mix of versions supported for 3 yrs and 2 yrs and we did not take that into account previously.
Revised the PR# + link, and slightly reworded the Aerospike addition to try to make it clearer that we have only added it to the list of recognized names.  Maybe of note - I think Java may also have some experimental support for Aerospike.