More Defensive zerologWriter parsing
In preparation for a more feature rich log data spec, this parser was
built to be able to scan additional contextual vaule from a zerolog
log message. This patch fixes a bug where it was unable to correctly
parse the JSON value for strings with commas,numbers, or stack traces.