Error Group Callbacks Work With High Security
When the agent is configured to protect user data from being ingested by
New Relic, the callback function should still have access to data that
would otherwise be scrubbed. The callback function is written by the
customer, and is authorized to handle sensitive data. This will not
change how sensitive data is sent to New Relic from the agent in any
way. Users are strongly advised not to include sensitive data in their
Error Group callback function.